requirements.txt
flask
flask-cors
torch
torchvision
numpy
opencv-python
Pillow
grad-cam
gunicorn  # <-- Add this line